backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.36
alembic==1.13.2
python-multipart==0.0.9
asyncpg==0.29.0
pydantic[email]==2.8.2
greenlet==3.0.2
psycopg[binary]>=3.1
psycopg2-binary